tests: fix unit tests
