+++ src/retry/tests/CircuitBreakerAnotationAspectTest.php
@@ -11,8 +11,6 @@ use Hyperf\Retry\Annotation\CircuitBreaker;
 use Hyperf\Retry\Aspect\RetryAnnotationAspect;
 use Hyperf\Retry\CircuitBreakerState;
 use Hyperf\Retry\FlatStrategy;
-use Hyperf\Retry\NoOpRetryBudget;
-use Hyperf\Retry\RetryBudgetInterface;
 use Hyperf\Utils\ApplicationContext;
 use HyperfTest\Retry\Stub\Foo;
 use Mockery;
@@ -30,6 +28,65 @@ class CircuitBreakerAnotationAspectTest extends TestCase
     protected function tearDown(): void
     {
         Mockery::close();
+    }
+
+    public function testCircuitBreaker()
+    {
+        $aspect = new RetryAnnotationAspect();
+        $point = Mockery::mock(ProceedingJoinPoint::class);
+
+        $point->shouldReceive('getAnnotationMetadata')->andReturns(
+            new class() extends AnnotationMetadata {
+                public $method;
+
+                public function __construct()
+                {
+                    $state = Mockery::mock(
+                        CircuitBreakerState::class
+                    );
+                    $state->shouldReceive('isOpen')->twice()->andReturns(false);
+                    $state->shouldReceive('isOpen')->once()->andReturns(true);
+                    $state->shouldReceive('open')->andReturns();
+                    $retry = new CircuitBreaker(['circuitBreakerState' => $state]);
+                    $retry->sleepStrategyClass = FlatStrategy::class;
+                    $this->method = [
+                        AbstractRetry::class => $retry,
+                    ];
+                }
+            }
+        );
+        $point->shouldReceive('process')->times(2)->andThrow(new \RuntimeException('ok'));
+        $point->shouldReceive('getArguments')->andReturns([]);
+        $this->expectException('RuntimeException');
+        $aspect->process($point);
+    }
+
+    public function testFallbackForCircuitBreaker()
+    {
+        $aspect = new RetryAnnotationAspect();
+        $point = Mockery::mock(ProceedingJoinPoint::class);
+
+        $point->shouldReceive('getAnnotationMetadata')->andReturns(
+            new class() extends AnnotationMetadata {
+                public $method;
+
+                public function __construct()
+                {
+                    $state = new CircuitBreakerState(10);
+                    $retry = new CircuitBreaker(['circuitBreakerState' => $state]);
+                    $retry->sleepStrategyClass = FlatStrategy::class;
+                    $retry->fallback = Foo::class . '@fallbackWithThrowable';
+                    $retry->maxAttempts = 2;
+                    $this->method = [
+                        AbstractRetry::class => $retry,
+                    ];
+                }
+            }
+        );
+        $point->shouldReceive('process')->times(2)->andThrow(new \RuntimeException('ok'));
+        $point->shouldReceive('getArguments')->andReturns([$string = uniqid()]);
+        $result = $aspect->process($point);
+        static::assertSame($string . ':ok', $result);
     }
 
     public function testFallbackWhenStateOpenFirst()
